Separate `CollapsibleCustomerCardAddressViewModel` to its own file.
